+++ src/MTG/Core/MyPHPMailer.php
@@ -1,12 +1,12 @@
 <?php
 
 /*
-Version:     1.12
+Version:     1.13
 Date:        11/01/26
 Name:        MyPHPMailer.php
 Purpose:     Extends PHPMailer with standard options.
 Notes:       Usage:
-                 $mail = new MyPHPMailer(true, $smtpParameters, $serverEmail, $logfile);
+                 $mail = new MyPHPMailer(true, $smtpParameters, $serverEmail, $logfile, $siteTitle, $appConfig);
                  $mailresult = $mail->sendEmail($adminEmail, false, $subject, $body);
 Author:      Simon Wilson
 Copyright:   2025 MTG Collection
@@ -33,16 +33,42 @@ class MyPHPMailer extends PHPMailer
     private $logfile;
     private $message;
     private $siteTitle;
+    private $emailEnabled;
+    private $appConfig;
 
-    public function __construct($exceptions, $smtpParameters, $serverEmail, $logfile, $siteTitle = null)
-    {
+    public function __construct(
+        $exceptions,
+        $smtpParameters,
+        $serverEmail,
+        $logfile,
+        $siteTitle = null,
+        ?AppConfig $appConfig = null
+    ) {
         //Don't forget to do this or other things may not be set correctly!
         parent::__construct($exceptions);
         // Set variables
+        $this->appConfig = $appConfig;
         $this->serverEmail = $serverEmail;
         $this->logfile = $logfile;
-        $this->message = new Message($this->logfile);
-        $this->siteTitle = $siteTitle ?: $GLOBALS['siteTitle'];
+
+        if ($this->appConfig !== null && ($this->serverEmail === null || $this->serverEmail === '')) :
+            $this->serverEmail = (string) $this->appConfig->email('serverEmail', '');
+        endif;
+        if ($this->appConfig !== null && $this->logfile === null) :
+            $this->logfile = $this->appConfig->general('logFile', null);
+        endif;
+
+        $logLevel = $this->appConfig ? $this->appConfig->general('logLevel', null) : null;
+        $this->message = new Message($this->logfile, $logLevel, $this->appConfig);
+        $this->siteTitle = $siteTitle
+            ?: ($this->appConfig ? (string) $this->appConfig->general('title', '') : ($GLOBALS['siteTitle'] ?? ''));
+        $this->emailEnabled = $this->appConfig
+            ? (bool) $this->appConfig->email('enabled', false)
+            : (isset($GLOBALS['emailEnabled']) && $GLOBALS['emailEnabled'] === true);
+
+        if (empty($smtpParameters) && $this->appConfig !== null) :
+            $smtpParameters = $this->appConfig->getSmtpParameters();
+        endif;
 
         // Set defaults for PHPMailer from ini.file
         $this->setFrom($this->serverEmail, $this->siteTitle);
@@ -93,7 +119,7 @@ class MyPHPMailer extends PHPMailer
         $attachmentname = '',
         $attachments = []
     ) {
-        if (!isset($GLOBALS['emailEnabled']) || $GLOBALS['emailEnabled'] !== true) :
+        if ($this->emailEnabled !== true) :
             $this->message->logMessage(
                 '[NOTICE]',
                 "Email disabled; skipping send to $recipient with subject '$subject'"